Fix doc comment
The `of` method for the `Cfg` struct mentions the command used to obtain
the configuration and parse. However, the command mentioned is
incorrect, so recreating/testing the command at a terminal prompt
results in an error. The correct command is added along with some
clarifying information about the value that should be passed as a
string.